feat: Refactor SQL utilities into @appland/models
